+++ src/plugins/plugin_14C/Plugin14CRefView.cpp
@@ -98,7 +98,7 @@ void Plugin14CRefView::setDate(const Date& d, const ProjectSettings& settings)
         
         for(int t=yMin; t<yMax; ++t)
         {
-            float v = exp(-0.5 * pow((age - t) / error, 2));
+            float v = expf(-0.5 * powf((age - t) / error, 2));
             curveMeasure.mData[t] = v;
         }
         curveMeasure.mData = normalize_map(curveMeasure.mData);
